+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 4.0.1-SNAPSHOT
+### 4.0.1
 - Sort fake chunks before loading so nearby chunks load first
 - Fix render distance resetting after restart if above 32 (#109)
 - Fix errors about Starlight in log when Starlight is not installed